==> notebooks/subnotebooks/02-format-courses.ipynb <==
# Format Courses

import pandas as pd

f = '../data/01-courses-result.json'
old_f = '../data/001-reference-courses.json'

df = pd.read_json(f)

print (df.columns)
df.head()

old_df = pd.read_json(old_f)
old_df.head()

old_df = old_df.transpose()

old_df.columns
old_df.rename(columns={'link': 'href', 'hrs': 'hours', 'prereqs': 'requirements'}, inplace=True)

get_link_src = lambda x: [i['source'] for i in x]
print (get_link_src([{'source': 'ECON:4800', 'target': 'ECON:6320'}]))

old_df['requirements'] = old_df['requirements'].apply(get_link_src)

old_df.reset_index(inplace=True, drop=True)
old_df.head()

# duplicates?
print (len(df))
print (len(df.courseID.unique()))

df.drop_duplicates(subset=['courseID'], inplace=True, keep='first')

### final attrs:

- [x] courseID
- [x] course num
- [x] department
- [x] formatted hours
- [x] href
- [x] description
- [ ] requirements list

form_department = lambda x: x.split(':')[0]
df['department'] = df['courseID'].apply(form_department)

form_course_num = lambda x: int(x.split(':')[1])
df['course_num'] = df['courseID'].apply(form_course_num)

def form_hrs(x):
    # get rid of the trailing ' s.h.'
    y = x[:-5]
    if (y != 'ARR'):
        try:
            return int(y.strip())
        except:
            z = y.split(',')[-1]
            return int(z.strip())
    else:
        return y
    
df['hours'] = df['hours'].apply(form_hrs)

from w3lib.html import replace_escape_chars
def form_html_text(x):
    unescaped = replace_escape_chars(x)
    single_spaced = ' '.join(unescaped.split())
    return single_spaced

df['description'] = df['description'].apply(form_html_text)

def get_req_candidates(x):
    if (x[0] is None):
        return []
    
    reqs = x[0]
    _id = x[1]
    
    ###
    # We could check if there's *any* hard course prereqs and leave the GPA, 
    # but thats for another day
    ###
    formatted = form_html_text(reqs)
    split = formatted.split()
    
    candidates = []
    for i in split:
        if ':' in i:
            i = i.replace(',', '')
            candidates.append(i)
            
    ###
    # Need to write logic for requirements for multiple classes
    # `CEE:4158—CEE:2150` => CEE:2150 is required for 4158, 
    # and we will need to only keep those with the appropriate courseID
    ###
    better_candidates = []
    for i in candidates:
        if '—' in i:
            split = i.split('—')
            if (_id == split[0]):
                better_candidates.append(split[1])
        else:
            better_candidates.append(i)
    
    if len(better_candidates) == 0:
        return []
    else:
        return list(set(better_candidates))
    
df['requirements'] = df[['requirements', 'courseID']].apply(get_req_candidates, axis=1)

df.head()

## Combine old and new DF's

old_slim = old_df[['courseID', 'description', 'hours', 'href', 'name', 'requirements', 'department', 'course_num']]
new_df = pd.concat([df, old_slim])

print (len(new_df))
new_df.drop_duplicates(subset=['courseID'], inplace=True)
print (len(new_df))

### Create links with valid nodes

# longer, more surefire way

# holds all our links
links_list = []

# define function that we'll use to create links
def create_course_links(x):
    if (x[0] is []):
        return
    
    reqs = x[0]
    _id = x[1]
    
    for req in reqs:
        link = {
            'source': req,
            'target': _id
        }
        links_list.append(link)
    
#df[['requirements', 'courseID']].apply(get_req_candidates, axis=1)
invalid_links = new_df[['requirements', 'courseID']].values.tolist()

for i in invalid_links:
    create_course_links(i)

valid_IDs = set(new_df['courseID'].unique().tolist())
valid_links = []

for link in links_list:
    src, trgt = link['source'], link['target']
    
    if src not in valid_IDs:
        print (link)
        continue
    elif trgt not in valid_IDs:
        print (link)
        continue
    else:
        valid_links.append(link)

print (len(links_list))
print (len(valid_links))

# create a map
before_links = {}
for i in valid_IDs:
    before_links[i] = []

# populate map
for i in links_list:
    if i['target'] in valid_IDs:
        if i['source'] in valid_IDs:
            course = before_links[i['target']]
            course.append(i)
    
# move the data to the correct row in DF
new_df['before'] = new_df['courseID'].apply(lambda x: before_links[x])

# create a map
after_links = {}
for i in valid_IDs:
    after_links[i] = []

# populate map
for i in links_list:
    if i['source'] in links_list:
        if i['target'] in valid_IDs:
            course = after_links[i['source']]
            course.append(i)
    
# move the data to the correct row in DF
new_df['after'] = new_df['courseID'].apply(lambda x: after_links[x])

### Sanity Check

See if theres any requirements in our dataframe that don't have an entry.

print (len(df) == len(df.courseID.unique()))

rendered_links = new_df['before'].values.tolist()

# [item for sublist in l for item in sublist]
rendered_links = [item for sublist in rendered_links for item in sublist]

for link in rendered_links:
    src, trgt = link['source'], link['target']
    
    if src not in valid_IDs:
        print (link)
        continue
    elif trgt not in valid_IDs:
        print (link)
        continue

***Nice***

Now just standardize the hours and be done!

new_df['hours'].unique()

new_df['hours'] = new_df['hours'].apply(lambda x: 0 if x is None else x)

#### At long last...

new_df.columns

new_df = new_df[['courseID', 'description', 'hours', 'href', 'name',
       'department', 'course_num', 'before', 'after']]

f_out = '../data/02-rowed-courses.json'
new_df.to_json(f_out, orient='records', lines=True)

f_out2 = '../data/02-indexed-courses.json'
new_df.set_index('courseID', drop=False)\
    .to_json(f_out2, orient='index')
