==> notebooks/subnotebooks/03-format-majors.ipynb <==
!dir

import pandas as pd

courses_f = '02-indexed-courses.json'
majors_f = 'academicplanresult.json'

courses = pd.read_json(courses_f).transpose()

majors = pd.read_json(majors_f)

### Format Columns

def format_major_name(dirty):
    split = dirty.split('-')
    n = len(split)
    
    for i in range(n):
        if (i == n-1):
            split[i] = split[i].upper()
        else:
            split[i] = split[i].title()
    
    return ' '.join(split)

majors['alias'] = majors['major'].apply(format_major_name)

majors.semester.unique()

year_map = {
    'First Year': 1,
    'Second Year': 2,
    'Third Year': 3,
    'Fourth Year': 4
}

semester_map = {
    'Fall': 1,
    'Spring': 2,
    'Summer': 3
}

majors['year'] = majors['year'].apply(lambda x: year_map[x])
majors['semester'] = majors['semester'].apply(lambda x: semester_map[x])

majors.head()

### Check if all major courses are valid

maj_records = majors.to_dict(orient='records')

valid_IDs = set(courses.courseID.unique())
invalid_IDs = []

for course in maj_records:
    courseID = course['courseID']
    maj = course['major']
    
    if courseID not in valid_IDs:
        print (courseID, maj)
        invalid_IDs.append(courseID)

len(invalid_IDs)

len(set(invalid_IDs))

corrections = {
    'PHYS:1702': 'PHYS:1701',
    'PHYS:2704': 'PHYS:2703',
    'PHYS:3742': 'PHYS:3741',
    'PHYS:2704': 'PHYS:2703',
    'PHYS:3812': 'PHYS:3811',
    
    'STAT:3101': 'STAT:3100',
    'STAT:3210': 'STAT:3200',
    
    'CHEM:2420': 'CHEM:2410',
    'CHEM:2240': 'CHEM:2230'
}



courses[courses.department == 'CHEM']

# make corrections before creating index
print (len(maj_records))

for course in maj_records:
    courseID = course['courseID']
    maj = course['major']
    
    if courseID not in valid_IDs:
        try:
            course['courseID'] = corrections[courseID]
        except:
            maj_records.remove(course)
            
print (len(maj_records))

### Create Indexed Data Structure

major_names = majors['major'].unique().tolist()

res = {}
for item in maj_records:
    cID = item['courseID']
    
    if cID in valid_IDs:
        res.setdefault(item['major'], []).append(item)

file_data = {}

create_slim = lambda x: {"courseID": x['courseID'], "semester": x['semester'], "year": x['year']}

for item in res:
    file_data[item] = {}
    
    file_data[item]['courses'] = [create_slim(i) for i in res[item]]
    file_data[item]['major'] = item
    file_data[item]['alias'] = format_major_name(item)
    
# remove any duplicates we may have
for item in file_data:
    #file_data[item]['courses'] = set(file_data[item]['courses'])
    seen_so_far = set()
    filt_courses = []
    
    for course in file_data[item]['courses']:
        if course['courseID'] not in seen_so_far:
            seen_so_far.add(course['courseID'])
            filt_courses.append(course)
    
    file_data[item]['courses'] = filt_courses

file_data

import json
with open('03-indexed-majors.json', 'w') as out:
    json.dump(file_data, out)
