==> notebooks/subnotebooks/04-department-format.ipynb <==
# Department Data

import pandas as pd

f = '../data/02-indexed-courses.json'

df = pd.read_json(f).transpose()

df.head()

departments = df['department'].unique()
departments.sort()

department_df = pd.DataFrame()
department_df['label'] = departments
department_df['value'] = departments

department_df.head()

### Get Count of Department Courses

dep_counts = df['department'].value_counts()

department_df['label'] = department_df['label'].apply(lambda x: x + ' (' + str(dep_counts[x]) + ')')

department_df.head()

#### Filter Rows with no value

#### Write out

out = '../data/04-departments.csv'
department_df.to_csv(out, index=False)
