==> notebooks/subnotebooks/05-class-construction.ipynb <==
// libraries
var fs = require('fs');
var http = require('http');

var _ = require('lodash');

var d3 = require('d3');
var d3_sankey = require('d3-sankey');

var jsdom = require('jsdom');

// files
var f1 = '../data/02-indexed-courses.json'; // new line delimited
var f2 = '../data/indexed-courses.json';

function read_json(f) {
    var buffer = fs.readFileSync(f);
    var contents = JSON.parse(buffer);
    return contents;
}

// Function to create and return an svg for our virtual DOM
function svgDOM(width, height) {
  // Setup DOM
  var document = new jsdom.JSDOM();
  var body = d3.select(document.window.document.body);
    
  // Create svg node
  return body.append('svg')
    .attr('xmlns', 'http://www.w3.org/2000/svg')
    .attr('xmlns:xlink', 'http://www.w3.org/1999/xlink')
    .attr('width', width)
    .attr('height', height);
}

var indexed_courses = read_json(f1);
var old_courses = read_json(f2);

class KnowledgeGraph:
	
	/* constructor
		attrs 
			knowledge, aka our known college courses
			
			nodes, data structure holding our permissible nodes
			links, data structure holding the links between our nodes
	*/
	
	/* load file
	*/
	
	/* add node
		*adds node to graph's permissible nodes*/
		// add node to permissible nodes
		node = this.knowledge[nodeID]
		? this.nodes.add(node)
		? this.nodes[nodeID] = node
		? this.nodes.push(nodeID|node)
	*/
	
	/* add link
		*adds a link to our links* both are nodeID
		link = {
			source: source,
			target: target,
			value: 1
		}
		this.links.push(links)
	*/
	
	
	/* is node permissible
		*fx to determine if a nodeID is permissible*
		return this.nodes.indexOf(nodeID) > -1
	*/
	
	
	/* clear graph nodes
		this.nodes = []
	*/
	
	
	/* clear graph links
		this.links = []
	*/
	
	
	/* update graph links
		*updates the links between courses in our graph*
		this.clear_graph_links()
		
		// add a nodes prereqs
		for nodeID in this.nodes:
			node = this.knowledge[nodeID]
			
			// add each permissible pre
			for prev in node.prev:
				if this.isPermissble(prev):
					this.addLink(prev, nodeID)
		
		// remove duplicates preemptively 
		this.links.removeDuplicates
	*/
	
	
	/* create graph from list
		// assuming course list of nodes input
		// start fresh
		this.clearNodes()
		this.clearLinks()
		
		// add all nodes, and then update links
		node_list.forEach(this.addNode)
		this.updateGraphLinks()
		
		// map the node IDs to actual nodes
		copy_nodes = this.nodes.map(
			copy = _.deepClone(this.knowledge[d])
			return copy
		)
		
		return {nodes: copy_nodes, links: this.links}
	*/
	
	/* create graph from search (nodeID)
		// start fresh
		this.clearNodes()
		this.clearLinks()
		
		// create queue to traverse graph
		queue = []
		queue.push(nodeID)
		
		// traversal
		while queue.length > 0:
			node = this.knowledge[queue.pop()]
			
			if node.prev.length > 0:
				node.prev.forEach(queue.push(d))
			
			// after to be implemented via data transform
			// add the antecedents
			
			this.addNode(node.nodeID)
		
		
	

class Knowledge:
	/* constructor
		this.courses = values.map(d.courseID => d)
	*/
	
	
	/* get subset (attractive, value)
		filtered_courses = d3.values(this.courses)
			.filter(return d[attr] === value)
			.map(d.courseID)
	*/
	



class Learning:
	/* constructor 
		*aka major/curriculum/academic plan*
	this.courses = [{nodeID, semester, year}]
	this.name = major
	this.alias = alias
	
	this.majors = {}
	
	*/
	
	/* add course
		* constructs and adds a course
		course = {
			nodeID: nodeID,
			semester: semester,
			year: year
		}
		this.courses.push(course)
	*/
	
	
	/* remove course
		this.courses = this.courses.filter(d.nodeID != nodeID)
	*/
	
	
	/* reset courses
		this.courses = []
	*/
	
	
	/* edit course
		this.name = "My Courses"
		// fill this
	*/
	
	/* assign major (majorID)
		major = this.majors[majorID]
		this.name = major.name
		this.alias = major.alias
		
		this.courses = major.courses
	*/
	
	
	/* get learning paths (Knowledge)
		// get list of course ids
		curriculum = this.courses.map(d.nodeID)
		
		// create graph
		kg = KnowledgeGraph(Knowledge)
		graph = kg.createGraphFromList(curriculum)
		
		// map semester and year back to nodes
		// to do
		
		return graph
	*/
	
### Knowledge (Repository)

function Knowledge(courses) {
    this.courses = courses;
    
    this.getSubset = function(attr, value) {
        var filtered_courses = d3.values(this.courses).filter(d => {
            return d[attr] === value;
        })
        .map(d => d.courseID);
        return filtered_courses;
    };
}

var test = new Knowledge(indexed_courses);
var test_filter = test.getSubset('department', 'ACCT');

### Knowledge Graph

Remember to fix the conditional adding of courses

and to fix adding the link syntax.

function KnowledgeGraph(knowledge) {
    // attrs
    this.knowledge = knowledge;
    this.nodes = [];
    this.links = [];
    
    
    this.addNode = function(nodeID) {
        this.nodes.push(nodeID);
    }
    
    this.addLink = function(link) {
        this.links.push(link);
    }
    
    this.isPermissible = function(nodeID) {
        return this.nodes.indexOf(nodeID) > -1;
    };
    
    this.dedupeNodes = function() {
        this.nodes = _.uniq(this.links, d => d.courseID);
        /*
        this.nodes.filter(function(value, index, self) { 
            return self.indexOf(value) === index;
        });
        */
    }
    
    this.dedupeLinks = function() {
        //_.uniq(array, element => element.name + element.age);
        this.links = _.uniq(this.links, d => d.source + d.target);
    }
    
    this.clearNodes = function() {
        this.nodes = [];
    }
    
    this.clearLinks = function() {
        this.links = [];
    }
    
    this.updateGraphLinks = function() {
        // save obj for later
        var kg = this;
        
        // Clear links
        this.clearLinks();
        
        // For each node, add it's permissible prereqs
        this.nodes.forEach(d => {
            var node = this.knowledge.courses[d];
            
            // FIX THIS
            // right now it checks node , shouldnt have to
            if (node !== undefined) {
              node['before']
                .filter(e => { return kg.isPermissible(e.source); })
                .forEach(e => { kg.addLink(e, d); });
            };
            //
            
        });
        
        // remove duplicate links
        this.dedupeLinks();
    }
    
    this.cloneGraphNodes = function() {
        return this.nodes.map(d => _.cloneDeep(this.knowledge.courses[d]));
    };
    
    this.createGraphFromList = function(node_list) {
        // obg for later
        var kg = this;
        
        // start fresh
        this.clearNodes();
        this.clearLinks();
        
        // add all nodes from list and then update links
        node_list.forEach(d => kg.addNode(d));
        this.dedupeNodes();
        
        this.updateGraphLinks();
        
        return {nodes: this.cloneGraphNodes(), links: this.links};
    };
    
    this.createGraphFromSearch = function(nodeID) {
        var kg = this;
        
        // start fresh
        this.clearNodes();
        this.clearLinks();
        
        // traversal queue
        var queue = [];
        queue.push(nodeID);
        
        // Search graph, adding nodes to our list along the way
        while (queue.length > 0) {
            var node = this.knowledge.courses[queue.pop()];
            
            if (node !== undefined) {
                // add node to our list
                kg.addNode(node.courseID);
                
                // add before
                node['before'].forEach(e => { 
                    // only push node if it's not permissible, AKA we haven't seen it before
                    if (kg.isPermissible(e['source']) === false) { queue.push(e['source']); }
                });
                
                // Add antecedents(after)
                // to-do
            };
        };
        
        // dedupe nodes before creating links
        this.dedupeNodes();
        this.updateGraphLinks();
        
        return {nodes: this.cloneGraphNodes(), links: this.links};
    }
    
}

var kg = new KnowledgeGraph(test);

var cs_classes = test.getSubset('department', 'CS');

var cs_graph = kg.createGraphFromList(cs_classes);

var algo_graph = kg.createGraphFromSearch('CS:3330');

algo_graph

***Fix Colleges in Format notebook***

var tippie_classes = test.getSubset('college', 'Tippie College of Business');
var tippie_graph = kg.createGraphFromList(tippie_classes);

var eng_classes = test.getSubset('college', 'College of Engineering');
var eng_graph = kg.createGraphFromList(eng_classes);

var departments = d3.set(d3.values(test.courses).map(d => d.department))

departments

eng_classes

function draw_chart(graph) {
    
    // sanity check
    if (graph.links.length < 1 || graph.nodes.length < 1) {
        console.log("No links, skipping.");
        return;
    }
    
    var width = 960,
        height = 600,
        svg = svgDOM(width, height);
    
    var departments = graph.nodes
        .map(d => { return d.department; })
        .filter(function(value, index, self) { 
            return self.indexOf(value) === index;
        });
    var color_scale = d3.scaleOrdinal(d3.schemeCategory10)
        .domain(departments);

    var sankey = d3_sankey.sankey()
        .nodeId(function(d) { return d.courseID; })
        .nodeAlign(d3_sankey.sankeyCenter)
        .nodeWidth(20)
        .nodePadding(15)
        .size([width, height]);
    
    sankey(graph);

    var link = svg.append("g")
        .attr("class", "links")
        .attr("fill", "none")
        .attr("stroke", "#000")
        .attr("stroke-opacity", 0.2)
      .selectAll("path");
    var node = svg.append("g")
        .attr("class", "nodes")
        .attr("font-family", "sans-serif")
        .attr("font-size", 10)
      .selectAll("g");

    link = link
        .data(graph.links)
        .enter().append("path")
        .attr("d", d3_sankey.sankeyLinkHorizontal())
        .attr('class', 'link')
        .attr("stroke-width", function(d) { return Math.max(1, d.width); });
    link.append("title")
        .text(function(d) { return d.source.name + " → " + d.target.name; });

    node = node
        .data(graph.nodes)
        .enter().append("g")
        .filter(function(d) { return (d['targetLinks'].length > 0) || (d['sourceLinks'].length > 0); })
        .attr('class', 'node');
    node.append("rect")
        .attr('id', d => { return d.courseID; })
        .attr('class', 'node')
        .attr("x", function(d) { return d.x0; })
        .attr("y", function(d) { return d.y0; })
        .attr("height", function(d) { return d.y1 - d.y0; })
        .attr("width", function(d) { return d.x1 - d.x0; })
        .attr("fill", function(d) { return color_scale(d.department); })
        .attr("stroke", "#000");
    var course_names = node.append("text")
        .attr("x", function(d) { return d.x0 - 6; })
        .attr("y", function(d) { return (d.y1 + d.y0) / 2; })
        .attr("dy", "0.35em")
        .attr('class', 'courseName')
        .attr("text-anchor", "end")
        .text(function(d) { return d.name; })
      .filter(function(d) { return d.x0 < width / 2; })
        .attr("x", function(d) { return d.x1 + 6; })
        .attr("text-anchor", "start");
    node.append("title")
        .text(function(d) { return d.name + "\n" + d.courseID +"\n" + d.department; });

    //return svg;
    return $$.svg(svg.node().outerHTML);
}

var search_test = kg.createGraphFromSearch('CS:2230');

search_test

draw_chart(search_test)
